Endrer tilbake til å ha hentDokument i egen rest client for å unngå trøbbel med tester.
