+++ src/main/java/no/nav/familie/integrasjoner/journalpost/JournalpostService.kt
@@ -1,5 +1,6 @@
 package no.nav.familie.integrasjoner.journalpost
 
+import no.nav.familie.integrasjoner.client.rest.SafHentDokumentRestClient
 import no.nav.familie.integrasjoner.client.rest.SafRestClient
 import no.nav.familie.kontrakter.felles.journalpost.Journalpost
 import no.nav.familie.kontrakter.felles.journalpost.JournalposterForBrukerRequest
@@ -7,7 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service
-class JournalpostService @Autowired constructor(private val safRestClient: SafRestClient) {
+class JournalpostService @Autowired constructor(private val safRestClient: SafRestClient,
+                                                private val safHentDokumentRestClient: SafHentDokumentRestClient) {
 
     fun hentSaksnummer(journalpostId: String): String? {
         val journalpost = safRestClient.hentJournalpost(journalpostId)
@@ -25,6 +27,6 @@ class JournalpostService @Autowired constructor(private val safRestClient: SafRe
     }
 
     fun hentDokument(journalpostId: String, dokumentInfoId: String, variantFormat: String): ByteArray {
-        return safRestClient.hentDokument(journalpostId, dokumentInfoId, variantFormat)
+        return safHentDokumentRestClient.hentDokument(journalpostId, dokumentInfoId, variantFormat)
     }
 }
